Extract ControlledJsonEditor into its own component file
